非登陆状态下，直接输入内部网址都要跳转到登录页面

+++ src/main/java/cn/hd/controller/ContractController.java
@@ -3,6 +3,7 @@ package cn.hd.controller;
 import cn.hd.model.BaseConditionVO;
 import cn.hd.model.Contract;
 import cn.hd.service.IContractService;
+import cn.hd.utils.CommonServiceUtil;
 import cn.hd.utils.StringUtil;
 import com.github.pagehelper.PageInfo;
 import org.springframework.stereotype.Controller;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpSession;
 
 
 @Controller
@@ -18,27 +20,40 @@ import javax.annotation.Resource;
 public class ContractController {
     @Resource
     IContractService contractService;
-
+    @Resource
+    CommonServiceUtil commonServiceUtil;
     @RequestMapping("/init")
-    public String init() {
+    public String init(HttpSession session) {
+        if (!commonServiceUtil.checkSession(session)) {
+            return "redirect:/login.jsp";
+        }
         return "contract/contract";
     }
 
     @RequestMapping("/delete/{id}")
-    public String delete(@PathVariable String id) {
+    public String delete(@PathVariable String id, HttpSession session) {
+        if (!commonServiceUtil.checkSession(session)) {
+            return "redirect:/login.jsp";
+        }
         contractService.deleteByPrimaryKey(id);
         return "redirect:/contract/query";
     }
 
     @RequestMapping("/updatePage/{id}")
-    public String updatePage(@PathVariable String id, Model model) {
+    public String updatePage(@PathVariable String id, Model model, HttpSession session) {
+        if (!commonServiceUtil.checkSession(session)) {
+            return "redirect:/login.jsp";
+        }
         Contract contract = contractService.selectByPrimaryKey(id);
         model.addAttribute("contract", contract);
         return "contract/contract";
     }
 
     @RequestMapping("/query")
-    public String query(BaseConditionVO vo, Model model) {
+    public String query(BaseConditionVO vo, Model model, HttpSession session) {
+        if (!commonServiceUtil.checkSession(session)) {
+            return "redirect:/login.jsp";
+        }
         PageInfo<Contract> pageInfo = contractService.query(vo);
         model.addAttribute("pageModel", pageInfo);
         model.addAttribute("vo", vo);
@@ -46,7 +61,10 @@ public class ContractController {
     }
 
     @RequestMapping("/save")
-    public String save(Contract contract) {
+    public String save(Contract contract, HttpSession session) {
+        if (!commonServiceUtil.checkSession(session)) {
+            return "redirect:/login.jsp";
+        }
         int k;
         if (StringUtil.isNotNull(contract.getId())) {
             k = contractService.updateByPrimaryKey(contract);
