+++ src/main/java/cn/hd/controller/DeptController.java
@@ -1,7 +1,9 @@
 package cn.hd.controller;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpSession;
 
+import cn.hd.utils.CommonServiceUtil;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,10 +24,13 @@ public class DeptController {
 
 	@Resource
 	IDeptService deptService;
-
+	@Resource
+	CommonServiceUtil commonServiceUtil;
 	@RequestMapping("/query")
-	public String query(BaseConditionVO vo, Model model) {
-		
+	public String query(BaseConditionVO vo, Model model, HttpSession session) {
+		if (!commonServiceUtil.checkSession(session)) {
+			return "redirect:/login.jsp";
+		}
 		PageInfo<Dept> pageInfo = deptService.query(vo);
 		model.addAttribute("pageModel", pageInfo);
 		model.addAttribute("vo", vo);
@@ -33,7 +38,10 @@ public class DeptController {
 	}
 
 	@RequestMapping("/save")
-	public String save(Dept dept) {
+	public String save(Dept dept, HttpSession session) {
+		if (!commonServiceUtil.checkSession(session)) {
+			return "redirect:/login.jsp";
+		}
 		if (StringUtil.isNotNull(dept.getDeptId())) {
 			int k = deptService.updateByPrimaryKey(dept);
 			System.out.println("**********" + k + "@@@@@@@@@@@@@@@");
@@ -50,22 +58,30 @@ public class DeptController {
 	}
 
 	@RequestMapping("/updatePage/{id}")
-	public String updatePage(@PathVariable String id, Model model) {
+	public String updatePage(@PathVariable String id, Model model, HttpSession session) {
+		if (!commonServiceUtil.checkSession(session)) {
+			return "redirect:/login.jsp";
+		}
 		Dept dept = deptService.selectByPrimaryKey(id);
 		model.addAttribute("dept", dept);
 		return "dept/dept";
 	}
 
 	@RequestMapping("/delete/{id}")
-	public String delete(@PathVariable String id) {
-
+	public String delete(@PathVariable String id, HttpSession session) {
+		if (!commonServiceUtil.checkSession(session)) {
+			return "redirect:/login.jsp";
+		}
 		deptService.deleteByPrimaryKey(id);
 
 		return "redirect:/dept/query";
 	}
 
 	@RequestMapping("/init")
-	public String init() {
+	public String init(HttpSession session) {
+		if (!commonServiceUtil.checkSession(session)) {
+			return "redirect:/login.jsp";
+		}
 		return "dept/dept";
 	}
